+++ src/test/java/domain/DeckTest.java
@@ -7,6 +7,7 @@ import domain.card.Card;
 import domain.card.CardNumber;
 import domain.card.CardShape;
 import domain.card.Cards;
+import domain.card.Deck;
 import domain.cardsGenerator.RandomCardsGenerator;
 import domain.participant.Dealer;
 import domain.participant.Player;
